decide to refactor productService
