+++ src/main/resources/templates/fragments/productsinstore.html
@@ -21,7 +21,7 @@
                         </div>
                         <div class="product-body">
                             <p class="product-category"><p th:text="${product.category.displayValue}"></p></p>
-                            <h3 class="product-name"><a href="#"><p th:text="${product.brand}"></p> <p th:text="${product.model}"></p></a></h3>
+                            <h3 class="product-name"><a href="product"><p th:text="${product.brand}"></p> <p th:text="${product.model}"></p></a></h3>
                             <h4 class="product-price"><p th:text="${product.price}"></p> <del class="product-old-price">$990.00</del></h4>
                             <div class="product-rating">
                                 <i class="fa fa-star"></i>
